requirements.txt
aiodns==3.0.0
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
click-plugins==1.1.1
colorama==0.4.6
filelock==3.12.4
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
pycares==4.3.0
pycparser==2.21
PyYAML==6.0.1
requests==2.31.0
requests-file==1.5.1
shodan==1.30.0
six==1.16.0
tldextract==3.6.0
urllib3==2.0.5
XlsxWriter==3.1.4
yarl==1.9.2
